crawler/requirements.txt
aiohttp
pymongo
pyyaml
beautifulsoup4
lxml
